Improved API Structure
Moved all the API Resources into their own submodule to keep the actual API code smaller and free
of all the Type Declarations for the Resources themselves as they are not as important in that case
